+++ index.html
@@ -236,7 +236,7 @@
   <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
   
   <!-- Подключаем тут скрипт -->
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
